Authorized.vue: convert to class components, modularize
