+++ src/views/Authorized/styles.scss
@@ -1,0 +1,81 @@
+@import '@/scss/_mixins.scss';
+
+@import '@/scss/wiggle.scss';
+
+#authorized {
+	$SideBar-width: calc(10rem + 10vw);
+
+	.blur-container {
+		width: 100%;
+		height: 100%;
+		background-color: $main-color;
+
+		&.blur-visible:not(#add-button):not(#new-student) {
+			filter: blur(1px);
+		}
+	}
+
+	#content {
+		height: 100%;
+		width: 100%;
+		#students {
+			padding: $students-top-padding 0rem 0rem $students-left-padding;
+			// height: 100%; // makes white footer
+			display: flex;
+			justify-content: flex-start;
+			align-items: flex-start;
+			align-content: flex-start;
+			box-sizing: border-box;
+
+			flex-direction: row;
+			flex-wrap: wrap;
+		}
+	}
+
+	div#cover {
+		position: absolute;
+		width: 100%;
+		height: 100%;
+		left: 0;
+		top: 0;
+		opacity: 0;
+		z-index: 10;
+		pointer-events: none;
+		background-color: black;
+		transition: opacity 300ms;
+
+		&.actually-hidden {
+			visibility: hidden !important;
+		}
+
+		&.visible {
+			visibility: visible;
+			opacity: 0.5;
+			pointer-events: all;
+		}
+	}
+
+	#nav-bar {
+		@include nav-shadow(1px);
+		@include nav-border(1px);
+		border-width: 0px;
+		border-right-width: 1px;
+		background-color: lighten($main-color, 2%);
+
+		width: $SideBar-width;
+		// flex-basis: $SideBar-width;
+		// height: 100vh; // with this too short
+		order: 0;
+	}
+	#nav-title {
+		@include nav-shadow(2px);
+		@include nav-border(1px);
+
+		border-width: 0px;
+		border-bottom-width: 1px;
+
+		background-color: lighten($main-color, 2%);
+
+		// height: calc(1vmin + 2rem);
+	}
+}
